+++ CHANGES.md
@@ -3,6 +3,7 @@
 0.6.0 plan: <https://github.com/greyli/apiflask/issues/31>
 
 - Support using the `output` decorator on async views.
+- Add `PaginationSchema` and `pagination_builder` as basic pagination support.
 
 
 ## Version 0.5.2